better reference frame acquisition and error mapping
