+++ acq4/devices/Microscope/Microscope.py
@@ -34,6 +34,7 @@ class Microscope(Device, OptomechDevice):
         self.currentSwitchPosition = None
         self.currentObjective = None
         self._focusDevice = None
+        self._positionDevice = None
         self._surfaceDepth = None
         
         self.objectives = collections.OrderedDict()
@@ -168,7 +169,7 @@ class Microscope(Device, OptomechDevice):
         """
         return self.mapToGlobal(QtGui.QVector3D(0, 0, 0)).z()
 
-    def setFocusDepth(self, z):
+    def setFocusDepth(self, z, speed='fast'):
         """Set the z-position of the focal plane.
 
         This method requires motorized focus control.
@@ -182,7 +183,7 @@ class Microscope(Device, OptomechDevice):
 
         # and this is where it needs to go
         fdpos[2] += dif
-        return fd.moveToGlobal(fdpos, 'fast')
+        return fd.moveToGlobal(fdpos, speed)
 
     def getSurfaceDepth(self):
         """Return the z-position of the sample surface as marked by the user.
@@ -193,6 +194,39 @@ class Microscope(Device, OptomechDevice):
         self._surfaceDepth = depth
         self.sigSurfaceDepthChanged.emit(depth)
         self.writeCalibration()
+
+    def globalPosition(self):
+        """Return the global position of the scope's center axis at the focal plane.
+        """
+        return self.mapToGlobal(QtGui.QVector3D(0, 0, 0))        
+
+    def setGlobalPosition(self, pos, speed='fast'):
+        """Move the microscope such that its center axis is at a specified global position.
+
+        If *pos* is a 3-element vector, then this method will also attempt to set the focus depth
+        accordingly.
+
+        Return a MoveFuture instance.
+
+        Note: If the xy positioning device is different from the z positioning
+        device, then the MoveFuture returned only corresponds to the xy motion.
+        """
+        pd = self.positionDevice()
+        fd = self.focusDevice()
+
+        if len(pos) == 3 and fd is not pd:
+            z = pos[2]
+            self.setFocusDepth(z)
+            pos = pos[:2]
+        if len(pos) == 2:
+            pos = list(pos) + [self.getFocusDepth()]
+
+        # Determine how to move the xy(z) stage to react the new center position
+        gpos = self.globalPosition()
+        sgpos = pd.globalPosition()
+        sgpos2 = pg.Vector(sgpos) + (pg.Vector(pos) - gpos)
+        sgpos2 = [sgpos2.x(), sgpos2.y(), sgpos2.z()]
+        return pd.moveToGlobal(sgpos2, speed)
 
     def writeCalibration(self):
         cal = {'surfaceDepth': self.getSurfaceDepth()}
@@ -207,6 +241,16 @@ class Microscope(Device, OptomechDevice):
                     break
                 p = p.parentDevice()
         return self._focusDevice
+
+    def positionDevice(self):
+        if self._positionDevice is None:
+            p = self
+            while True:
+                if p is None or isinstance(p, Stage) and p.capabilities()['setPos'][0] and p.capabilities()['setPos'][1]:
+                    self._positionDevice = p
+                    break
+                p = p.parentDevice()
+        return self._positionDevice
 
 
 class Objective(OptomechDevice):
